#505 Remove GroovyUnitTest class and UnitTest
