+++ ratpack-core/src/main/java/ratpack/parse/ParserSupport.java
@@ -35,7 +35,6 @@ import com.google.common.reflect.TypeToken;
  *
  * import java.io.UnsupportedEncodingException;
  *
- * import ratpack.test.UnitTest;
  * import ratpack.test.handling.HandlingResult;
  * import ratpack.test.handling.RequestFixture;
  *
@@ -83,7 +82,7 @@ import com.google.common.reflect.TypeToken;
  *
  *   // unit test
  *   public static void main(String[] args) throws Exception {
- *     HandlingResult result = UnitTest.handle(new ToUpperCaseHandler(), new Action&lt;RequestFixture&gt;() {
+ *     HandlingResult result = RequestFixture.handle(new ToUpperCaseHandler(), new Action&lt;RequestFixture&gt;() {
  *       public void execute(RequestFixture fixture) throws Exception {
  *         fixture
  *           .body("123456", "text/plain")
