+++ ratpack-core/src/test/groovy/ratpack/stream/tck/PeriodicPublisherVerification.groovy
@@ -21,6 +21,7 @@ import org.reactivestreams.tck.PublisherVerification
 import org.reactivestreams.tck.TestEnvironment
 import ratpack.stream.Streams
 
+import java.time.Duration
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledExecutorService
